petcare/Pawtaker-Backend-master/requirements.txt
bcrypt==4.0.1
black==23.1.0
Django==4.1.6
django-cors-headers==3.14.0
django-extensions==3.2.1
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.1
drf-spectacular-sidecar==2023.3.1
gunicorn==20.1.0
Pillow==9.4.0
psycopg2==2.9.5
PyJWT==1.7.1
python-dotenv==1.0.0
requests==2.28.2
phonenumbers==8.13.7
django-phonenumber-field==7.0.2
django-email-log==1.3.0
sentry_sdk==1.29.2
django-ckeditor==6.7.0
